refactor for 3.0
