+++ bnlp/__init__.py
@@ -1,15 +1,16 @@
-__version__="2.0.0"
+__version__="3.0"
 
 
 import os
-from bnlp.sentencepiece_tokenizer import SP_Tokenizer
-from bnlp.nltk_tokenizer import NLTK_Tokenizer
-from bnlp.basic_tokenizer import BasicTokenizer
-from bnlp.bengali_word2vec import Bengali_Word2Vec
-from bnlp.bengali_fasttext import Bengali_Fasttext
-from bnlp.glove_wordvector import BN_Glove
 from bnlp.pos import POS
 from bnlp.ner import NER
+from bnlp.tokenizer.nltk import NLTKTokenizer
+from bnlp.tokenizer.basic import BasicTokenizer
+from bnlp.tokenizer.sentencepiece import SentencepieceTokenizer
+from bnlp.embedding.word2vec import BengaliWord2Vec
+from bnlp.embedding.fasttext import BengaliFasttext
+from bnlp.embedding.glove import BengaliGlove
+
